FWsystemIN
My new <basic> firewall frame work is in and seems to be working. I have QA'ed and tested on RH/OeL and centos firewalld. Right now it still just adds the ports to the firewalld system if enabled via a firewalld created services .xml file.

I want to add a submenu for for firewall control functions.